fastapi_service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
openai==1.12.0
pdfplumber==0.10.3
python-docx==1.1.0
pytesseract==0.3.10
Pillow==10.2.0
python-multipart==0.0.6
python-dotenv==1.0.0
langchain==0.1.9
langchain-community==0.0.24
langchain-text-splitters==0.0.1
faiss-cpu==1.7.4
tiktoken==0.6.0
